Compare prefixed dataset IDs correctly.
Move '_dataset_ids_equal' from 'helpers' -> 'key', because it is only
useful for keys, and we don't want the cycle.
