+++ gcloud/datastore/test_key.py
@@ -247,6 +247,16 @@ class TestKey(unittest2.TestCase):
         self.assertFalse(key1 == key2)
         self.assertTrue(key1 != key2)
 
+    def test___eq_____ne___same_kind_and_id_different_dataset_pfx(self):
+        _DATASET = 'DATASET'
+        _DATASET_W_PFX = 's~DATASET'
+        _KIND = 'KIND'
+        _ID = 1234
+        key1 = self._makeOne(_KIND, _ID, dataset_id=_DATASET)
+        key2 = self._makeOne(_KIND, _ID, dataset_id=_DATASET_W_PFX)
+        self.assertTrue(key1 == key2)
+        self.assertFalse(key1 != key2)
+
     def test___eq_____ne___same_kind_different_names(self):
         _DATASET = 'DATASET'
         _KIND = 'KIND'
@@ -288,6 +298,16 @@ class TestKey(unittest2.TestCase):
                              namespace=_NAMESPACE2)
         self.assertFalse(key1 == key2)
         self.assertTrue(key1 != key2)
+
+    def test___eq_____ne___same_kind_and_name_different_dataset_pfx(self):
+        _DATASET = 'DATASET'
+        _DATASET_W_PFX = 's~DATASET'
+        _KIND = 'KIND'
+        _NAME = 'one'
+        key1 = self._makeOne(_KIND, _NAME, dataset_id=_DATASET)
+        key2 = self._makeOne(_KIND, _NAME, dataset_id=_DATASET_W_PFX)
+        self.assertTrue(key1 == key2)
+        self.assertFalse(key1 != key2)
 
     def test___hash___incomplete(self):
         _DATASET = 'DATASET'
@@ -453,3 +473,27 @@ class TestKey(unittest2.TestCase):
         self.assertEqual(parent.path, _PARENT_PATH)
         new_parent = key.parent
         self.assertTrue(parent is new_parent)
+
+
+class Test__dataset_ids_equal(unittest2.TestCase):
+
+    def _callFUT(self, dataset_id1, dataset_id2):
+        from gcloud.datastore.key import _dataset_ids_equal
+        return _dataset_ids_equal(dataset_id1, dataset_id2)
+
+    def test_identical_prefixed(self):
+        self.assertTrue(self._callFUT('s~foo', 's~foo'))
+        self.assertTrue(self._callFUT('e~bar', 'e~bar'))
+
+    def test_different_prefixed(self):
+        self.assertFalse(self._callFUT('s~foo', 's~bar'))
+        self.assertFalse(self._callFUT('s~foo', 'e~foo'))
+
+    def test_all_unprefixed(self):
+        self.assertTrue(self._callFUT('foo', 'foo'))
+        self.assertFalse(self._callFUT('foo', 'bar'))
+
+    def test_unprefixed_with_prefixed(self):
+        self.assertTrue(self._callFUT('foo', 's~foo'))
+        self.assertTrue(self._callFUT('foo', 'e~foo'))
+        self.assertFalse(self._callFUT('foo', 's~bar'))
